drf_clientes/requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.0.8
djangorestframework==3.13.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==3.0.0a5
pytz==2020.1
six==1.15.0
sqlparse==0.4.2
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1